requirements.txt
tensorflow>=2.20.0
streamlit>=1.52.0
pillow>=12.0.0
numpy>=2.0.0
